move spritesheet to 0x1000
the goal was to use 0x0-0x1000 for memory, but then I tested
things and re-realized why that isn't faster:

  prof(function(x,y,mem)
    mem[x]=y --2
  end,function(x,y,mem)
    poke4(x,y) --6
  end,function(x,y,mem)
    return $x --1
  end,function(x,y,mem)
    return mem[x] --2
  end,{locals={1,13,{14}}})

reading is 2x faster from actual memory, but writing is 3x slower,
so just leave module memory as a memory table.

but I did make some adjustments (token savings, flip the sprite shown
for tracker v. patch mode) so I'm leaving this commit. maybe we'll
find a use for the cleared memory at some point in the future